requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
Django==2.1.4
django-extensions==2.1.4
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
Fabric==1.14.1
foursquare==1!2016.9.12
idna==2.7
invoke==1.2.0
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
openapi-codec==1.3.2
paramiko==2.4.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
python-memcached==1.59
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
